refactor(sync-engine+query+core): right-size the engine door
The door exports the values it was starving consumers of
(SYNC_COLLECTION_NAMES, MUTATION_QUEUE_RXDB_COLLECTION — the hand-copied
9-name lists collapse to one import) and drops three zero-caller handle
members that merely duplicated surviving surface (stats ⊂ status,
onScopeEvent ⊂ events, censusTotals ⊂ censusChanges). conflicts()/
resolveConflict() stay: zero callers too, but they are the only surface
for terminal conflict rows — capability, not duplication.

packages/query's engine-monitor.ts deletes (a literal pass-through, a
byte-identical copied list, a shadowed type); its two count observers
move to their one consumer in core. The 12-type re-export block leaves
packages/query — core imports engine types from the engine.

ADR 0018's package-exports pin is amended deliberately: still two doors,
still a strict exact-list assertion — the list now has three runtime
values instead of one.
