+++ packages/core/src/screens/main/health/database-logic.test.ts
@@ -1,4 +1,4 @@
-import type { CensusTotals } from '@wcpos/query';
+import type { CensusTotals } from '@wcpos/sync-engine';
 
 import {
 	censusFreshnessWindow,
